UNIX domain socket support for Clozure CL.
As with SBCL, this adds support for UNIX domain sockets to Clozure CL.
For Darwin-based machines (i.e. all Macs), we also explicitly add
support for the correct default UNIX socket directory.

The code that constructs the UNIX socket path has been abstracted out
too, since we use it in multiple places now.

It seems to me that INITIATE-CONNECTION is beginning to get a bit
unwieldy, and could probably benefit from restructuring, but I have
intentionally avoided that to reduce the impact of this change.