As per postgresql's convention if the conection host is a string which starts with #\/ connect to a unix socket in that diretory.
To simplify some of the #+ccl and #+sbcl and avoid some code duplication
usocket is always used when connecting to an inet socket.

The change only affects those connections whose host is a string or a
pathname, :unix as a host still works as before (and depends on the
conection time value of *unix-socket-dir*).