Optimize WITH-CONNECTION for quoted lists
Because of https://bugs.launchpad.net/sbcl/+bug/1750466 the `APPLY` call in `WITH-CONNECTION` generates compiler notes in SBCL.

`SPEC` is evaluated. Because `SPEC` might be a variable or a function call that produces the spec, it is important to keep the call to `APPLY`, but we can typecheck at macroexpansion time to see if we are dealing with a quoted list and optimize for that scenario.

Therefore, if the macro is called like `(with-connection '("foo" "bar" ...) ...)`, the `APPLY` call, unnecessary in this scenario, is replaced with a `FUNCALL`.

This is a small optimization that breaks the clarity of the `WITH-CONNECTION` macro, therefore please only merge it if you think that it is worth it.