Add sql-ops for VALUES, GROUPING-SETS, and empty-set
Also correct filter and order-by parameters in STRING-AGG

I would note that the empty-set operator is a kludge, returning a list
containing a string of two parentheses as an empty set.

IMPORTANT: The addition of the VALUES sql-op may be a breaking change
if you already use VALUES outside of an INSERT-INTO statement in a simple
way with the automatic code generation. See tests.lisp for examples.

Grouping-sets allow multiple group-by in a single query (more
efficient than multiple select calls joined by UNION.