Revert issue 70 fix for more rethinking.
Issue 70 requests that dao-slots be allowed to have names
that are reserved words. This would violate postgresql column
requirements. Query whether it makes any sense to have dao column
name that cannot be used as a table column name. Or I am not
understanding the issue sufficiently. Temporary reversion is
best course at the moment.