DEF-SQL-OP: Use Alexandria's parse-body
Checking if is more the CAR of the body is a documentation
string is not enough to determine if we are looking at a docstring. If
the string is the only element of the body then it is part of the method
body. Leverage ALEXANDRIA:PARSE-BODY to extract the documentation string and
any declarations in DEF-SQL-OP.

Another defect on DEF-SQL-OP's implementation is the Unwanted variable
capture of OP. Gensym it.