Bug fixes for issues 178, 179 on prepared statements
Upon connecting to a postgresql database, postmodern puts various parameters
into a hash table in the parameters slot in a database-connection class
instance. With this commit, postmodern also collects the postgresql pid for
a session connection and places that in the in the parameters slot hashtable
as well. It is accessible either using postmodern:get-pid or via
cl-postgres:connection-pid which will return a list of the pid and the session
"secret-key". The secret-key is potentially useful for cancelling queries, but
that functionality is not yet included in postmodern.

Postmodern has added conditions for:
- database-error-invalid-sql-statement-name
- database-error-syntax-error-or-access-violation

There is substantial additional functionality regarding prepared statements. It
should be noted that prepared statements exist for the life of a connection and
are stored both in postgresql and in the meta slot of the postmodern
database-connection class.

It has been pointed out that from time to time over
a network connection, a prepared statement may be created and stored in the
postmodern database-connection class instance, but due to network failure,
timeout, etc has not been provided to postgresql, generating an
invalid-sql-statement-name error when an application tries to use a prepared
statement. Attempts to cancel the prepared statement query while maintaining the
original connection are often rejected. With this commit and in this type of
situation, postmodern will attempt to terminate the current session and
reconnect, resetting the prepared statement in postgresql.

This commit also adds the ability to specify prepared statement names that will
be used by postgresql using the defprepared macro. This can help in reading
postgresql log files. Prepared statements generated by the prepare macro will
still use the automatically generated statements.

With this commit, any generation of a prepared statement with the same name as an
existing prepared statement will overwrite the existing statement, effectively
allowing you to edit existing statements. Feedback is requested whether this
should be optional. Please note that since prepared statements using the prepare
macro have automatically generated names, this only impacts prepared statements
using the defprepared macro.

The following new functions are exported:

- (prepared-statement-exists-p (name))
This returns t if the prepared statement exists in the current
postgresql session, otherwise nil.

- (list-prepared-statements)
This is syntactic sugar. It runs a query that lists the prepared statements in
the session in which the function is run.

- (drop-prepared-statement (statement-name &key (location :both) (database *database*))
Prepared statements are stored both in the meta slot in the postmodern
connection and in postgresql session information. If you know the prepared
statement name, you can delete the prepared statement from both locations (the
default behavior), just from postmodern (passing :postmodern to the location
key parameter) or just from postgresql (passing :postgresql to the location
key parameter). If you pass the name 'All' as the statement name, it will
delete all prepared statements.

- (list-postmodern-prepared-statements (&optional (names-only nil))
List the prepared statements that postmodern has put in the meta slot in
the connection. It will return a list of alists of form:
  ((:NAME . \"SNY24\")
  (:STATEMENT . \"(SELECT name, salary FROM employee WHERE (city = $1))\")
  (:PREPARE-TIME . #<TIMESTAMP 25-11-2018T15:36:43,385>)
  (:PARAMETER-TYPES . \"{text}\") (:FROM-SQL)

If the names-only parameter is set to t, it will only return a list of
the names of the prepared statements.

- (find-postgresql-prepared-statement (name))
Returns the specified named prepared statement (if any) that postgresql has for this
session.

- (find-postmodern-prepared-statement (name))

Returns the specified named prepared statement (if any) that postmodern has
put in the meta slot in the connection. Note that this is the statement itself,
not the name.

- (reset-prepared-statement) (condition)
  If you have received an invalid-prepared-statement error but the prepared
statement is still in the meta slot in the postmodern connection,
try to regenerate the prepared statement at the database connection level.

- (overwrite-prepared-statement) (condition)
  If you have received an invalid-prepared-statement error but the prepared
statement is still in the meta slot in the postmodern connection,
try to regenerate the prepared statement at the database connection level.

- (get-pid)
Get the process id used by postgresql for this connection.

- (get-pid-from-postmodern)
Get the process id used by postgresql for this connection,
but get it from the postmodern connection parameters.

There are also two additional functions to cancel or terminate a postgresql session

- (cancel-backend)
Polite way of terminating a query at the database (as opposed to calling close-database).
Slower than (terminate-backend pid) and does not always work.

- (terminate-backend)
Less polite way of terminating at the database (as opposed to calling close-database).
Faster than (cancel-backend pid) and more reliable.