Optionality available for overwriting prepared statements
When *allow-overwriting-prepared-statements* is set to t (the default)
Prepared statements created with defprepared will overwrite
existing prepared statements of the same name if the query statement
provided to defprepared is different than the query statement
which exists in the postmodern meta connection hashtable. (remember
the prepared statements are stored separately for each postgresql
session). When set to nil, existing prepared statements with the
same name will not be overwritten regardless of whether the statements
are the same.

This parameter is exported.