Resolve Issue 70
Allow column names to be reserved words. Per
https://www.postgresql.org/docs/current/sql-keywords-appendix.html
postgresql reserved words can be used as column labels. A very old
version of postmodern allowed this, but the functionality disappeared
around 2014. The functionality has been restored, but there are some
important caveats:

Per the documentation:

"Labeled “reserved” are those tokens that are not allowed as column
or table names. Some reserved key words are allowable as names for
functions or data types; this is also shown in the table. If not so
marked, a reserved key word is only allowed as an “AS” column label
name."

It is possible, as demonstrated in this commit, to quote reserved
words and postgresql will accept that at the moment. There is no
guarantee that this will continue to work. If you are not using
reserved words as column names, this commit should not impact you
at all.