Defprepared overwrite ability can be toggled.
Set pomo:*allow-overwriting-prepared-statements* to t
to turn on defprepared's ability to overwrite older
prepared statements with the same name. Set it to nil
to turn off the overwrite ability.

If the overwrite ability is set, a named prepared statement
generated by defprepare will still only overwrite (if necessary)
the first time it is funcalled. Subsequent funcall's will
not trigger the check to see if overwriting an older prepared
statement is necessary.