Add file attributes for Genera
Genera requires file attributes to be set for proper
compilation. These are set by the first line of the file.  Briefly,
the attribute line sets certain properties of the file when it is
read, and these are used when compiling and loading. For example
'Syntax' specifies one of the (at least) three dialects of LISP that
Genera runs:

LISP -> ZetaLisp
COMMON-LISP -> CLtL2
ANSI-COMMON-LISP -> ANSI LISP

A common problem is people putting COMMON-LISP (CltL2) in for
Syntax. It's close enough that it will sometimes work, but often
introduces subtle bugs. This is probably partly because emacs also
uses this line to set some editing modes and people do not realize
that they have real meaning to some compilers and are not just
cosmetic.

The Package attribute is equivalent to (in-package). Genera
honors (in-package), but not until after compiling and loading the
file using the Package attribute. If the Package attribute wasn't set,
it uses the current package, which is often the last file compiled, so
you end up with all the definitions being loaded into packages off by
one. There is an open bug for this, as the ANSI spec says compile &
load should rebind package & read-table to the values they had before
processing the file.