s-sql drop-table can now use strings as well as symbols
Resolves reopened issue 167 that the s-sql operator :create-table can
use symbols or strings as the name of the table, but the s-sql :drop
operators could only use symbols. This restores the symmetry, but use
of symbols rather than strings is still recommended.

There is still an open question regarding :create-table xyz where the
table name is neither a string nor a symbol but currently is still
accepted. This does preclude using variables however. Is it worth
breaking some historic code to provide for the use of variables in the
table name or does anyone care.