Additional port flexibility to testing
Yes, with-connect in postmodern has a different parameter signature
(keywords) than with-connect in cl-postgres (just optional
parameters). Historical and changing it would break downstream
libraries. So, yes, ugly hack to provide flexibility in testing for
different ports and ssl configurations but just for testing purposes
and does not effect the base packages.