Fix unpadded octet-array and extra calls to make-prng
Postgresql's scram-sha-256 authentication requires a 32 element octet
array. Logxor against the client-key and client-signature returned an
integer that when converted to octets had a shorter array if the
client-key and client-signature had matching entries. That would
trigger a protocol violation when attempting scram-sha-256
authentication with postgresql. Padding the array resolved that
problem.

The previous code also made unnecessary calls to make-prng, slowing
down the process considerably.