Added functions rename-table, rename-column
function rename-table (old-name new-name)

Renames a table. Parameters can be strings or symbols. If you are renaming
a table using a fully qualified schema.table-name, you do not need to
specify the schema in the new-name. You cannot use this function to move
tables from one schema to another. Returns t if successful.

function rename-column (table-name old-name new-name)

Rename a column in a table. Parameters can be strings or symbols. If
the table is not in the public schema, it needs to be fully qualified
- e.g. schema.table. Returns t if successful.

Also added more tests for return types as well
as tests for rename-table and rename-column