Correct typo in postmodern test files
with-test-connection was in the wrong place in several tests causing
tests to fail with no database connection in test-dao.lisp and
prematurely closed in postmodern/tests/tests.lisp

Corrected timestamp tests for different timezones

The prior version of the tests only worked if the computer used as the
test platform (not the database server) was in UTC-5 timezone. This
correction temporarily resets local-time:*default-timezone* to UTC so
that timezone testing is consistent regardless of where in the world
the tester is located.