Pausing point while we regroup
Assume you pass three parameters, one an integer, one is text and the
third is something like a uuid or point or anything else that we do
not have a 1:1 correspondence between common lisp types and Postgresql
types. If you pass uuid as a text parameter, you will throw a
Postgresql error for type mismatch. If you do not specify any
parameter types, then the integer might come back as a string.

Also currently have a problem with negative floats and double floats
as well.