Flagging potential man-in-the-middle attack
Postgresql discovered a potential security issue in its server and
standard driver which might allow a potential man-in-the-middle attack
while attempting to make an ssl connection. See
[https://www.postgresql.org/about/news/postgresql-141-135-129-1114-1019-and-9624-released-2349/](https://www.postgresql.org/about/news/postgresql-141-135-129-1114-1019-and-9624-released-2349/)
for more details. Whilst Postmodern does not have this issue, after
discussions  with the Postgresql security team, we agreed to flag
evidence of attempted attacks using this vector.

If Postmodern is trying to set up an ssl connection and
unexpected extra data was received prior to an ssl connection being
created, such unexpected extra data may indicate an attempted man-in-the-middle
attack. With this commit, in such a situation, Postmodern will check
the new cl-postgres exported
variable *on-evidence-of-man-in-the-middle-attack* to its response. By
default, this variable is set to :error. You can set the
response to be a simple warning (by setting it to :warn) or you can
set it to :ignore.