Drop-role now works even if no postgres user
In order to drop a role, you need to reassign ownership of any
database objects it might own. Normal practice has been to reassign
ownership to a specified owner or to the default base role 'postgres'.

Some cloud setups do not give access to the postgres role, so it
cannot be used for this purpose. This change will assign the ownership
rights to the role calling drop-role if the new-owner parameter does
not actually exist.