Dollar Quoted tags restrictions relaxed
Dollar quoted tags are allowed in sql files. Prior to Postmodern
version 1.33.8 only alphabetic characters were allowed in tags in
order to use the execute-file function. Postgresql documentation and
industry practice allow any character to be in a dollar quoted
tag. Postmodern version 1.33.8 relaxes the alphabetic character
requirement. The only limitation now is that digit characters cannot
be in the first position in a tag.