Fix macro postmodern:defprepared which wasn't threadsafe when used with binary connections
When used with binary connections this macro generated code which
calls the `find-postmodern-prepared-statement` function and when called in the second thread,
this function returns `NIL`, because there is another connection in the `*database*` variable
and it has empty `(postmodern::connection-meta *database*)` hash.

Also, this commit fixed the problem with `mismatched-parameter-types` condition
which was signaled with unbound slots because it's initargs were specified as usual symbols
instead of keywords.