Handle no keys in the dao class.
If I make a dao class with no keys, such as

```
(postmodern:define-dao-class test-no-keys ()
  ((current boolean :accessor get-current :initarg :current)
   (value text :accessor get-value :initarg :value)))
```

and make an instance of it, I get an error

```
The value NIL is not of type STRING when binding STRING
```

This is caused by `(defmethod finalize-inheritance :after ((class dao-class))` that does

```
 (when (not (dao-keys class))
    (setf (slot-value class 'effective-keys)
          (list (find-primary-key-column class))))
```

so the `effective-keys` is `(NIL)`.

Now I am not sure this combination is documented as not allowed, but if so, definitely better error message would be useful, but setting the effective keys to `nil` (not as list) would be nice.

If you wonder why I create a class without any key, I use it to query-dao views to it.